+++ src/main/resources/static/js/standard.js
@@ -67,8 +67,11 @@ $(function () {
         });
     }
 
-    $("#tbody").on("click",".button_delete",function (){
-        deleteFault($(this));
+    $("#tbody").on("click",".button_delete",function () {
+        var r = confirm("是否删除?");
+        if (r == true) {
+            deleteFault($(this));
+        }
     });
 
     $("#tbody").on("click",".button_change",function () {
